Fixed bug in menu class. Adjusted disabled menuItem color.
Added search for first enabled menuItem that should be selected in
bounds of the items list. Removed automatic selection of prefered item,
even if it's disabled.

Adjusted disabled menu item color to darker grey. The gothic grey wasn't
that grey that it would be considered as disabled - greyed out item.